Port code changes from incentive (06/2023)

+++ layer1/Ray.cpp
@@ -4063,7 +4063,7 @@ int RayTraceThread(CRayThreadInfo * T)
                           dotgle = _1;
                         else if(dotgle < _0)
                           dotgle = _0;
-                        dotgle = powf(dotgle, 0.29F);
+                        dotgle = pow(dotgle, 0.29F);
                       } else {
                         dotgle = -dot_product3f(r1.surfnormal, bp->SpecNormal); /* fast OpenGL-like global specular */
                       }
